Update README.md
As an optional note:

At the bottom of the HTML page are a series of click events on the horizontal list of square color-coded boxes.

The click event design was for the internet location of the script to run.

I have to change the click event to the local host server for the event to work properly. Ouch! I will fix this soon.